fix ub caused by c_str() on temporary std::string
Using `c_str()` on a temporary `std::string` is undefined behavior. This
is because `c_str()` returns a pointer to the underlying memory location
which is only valid for the lifetime of the string itself. When used on
a temporary `std::string`, the value is release immediately after
calling `c_str()`, so the pointer is immediately invalidated.

Refs: https://stackoverflow.com/questions/35980664/why-does-calling-stdstring-c-str-on-a-function-that-returns-a-string-not-wor
PR-URL: https://github.com/ibmruntimes/node-zcrypto/pull/2
Reviewed-By: Gaby Baghdadi <baghdadi@ca.ibm.com>
Reviewed-By: Dhrumil Rana <dhrumil.rana@ibm.com>
Reviewed-By: Alex Yung <alex.yung@ibm.com>